+++ Blog/Blog-manager/pom.xml
@@ -3,7 +3,7 @@
     <modelVersion>4.0.0</modelVersion>
     <groupId>com.heleeos.blog</groupId>
     <artifactId>blog-manager</artifactId>
-    <version>1.0.0</version>
+    <version>1.0.1</version>
     <packaging>war</packaging>
 
     <dependencies>
