+++ Blog/Blog-manager/src/main/resources/applicationContext-database.xml
@@ -6,9 +6,9 @@
     
     <!-- 连接业务数据库 -->
     <bean id="dataSource" class="com.alibaba.druid.pool.DruidDataSource">
-        <property name="url" value="jdbc:mysql://127.0.0.13306/heleeos_blog" />
-        <property name="username" value="blog" />
-        <property name="password" value="blog123" />
+        <property name="url" value="jdbc:mysql://127.0.0.1:3306/heleeos_blog" />
+        <property name="username" value="root" />
+        <property name="password" value="admin" />
         <property name="initialSize" value="1" />
         <property name="minIdle" value="1" />
         <property name="maxActive" value="20" />
@@ -19,12 +19,13 @@
 
     <bean id="sqlSessionFactory" class="org.mybatis.spring.SqlSessionFactoryBean">
         <property name="dataSource" ref="dataSource" />
-        <property name="mapperLocations" value="classpath:com/heleeos/blog/**/mapper/*.xml" />
+        <property name="configLocation" value="classpath:mybatis-config.xml" />
+        <property name="mapperLocations" value="classpath:mappers/*.xml" />
     </bean>
     
     <bean id="scannerConfigurer" class="org.mybatis.spring.mapper.MapperScannerConfigurer">
         <property name="sqlSessionFactoryBeanName" value="sqlSessionFactory" />
-        <property name="basePackage" value="com.heleeos.blog.**.mapper" />
+        <property name="basePackage" value="com.heleeos.blog.dao" />
     </bean>
 
     <bean id="transactionManager" class="org.springframework.jdbc.datasource.DataSourceTransactionManager">
