+++ Blog/Blog-manager/src/main/resources/log4j.properties
@@ -1,0 +1,8 @@
+log4j.rootCategory=ERROR, stdout
+
+log4j.appender.stdout=org.apache.log4j.ConsoleAppender
+log4j.appender.stdout.layout=org.apache.log4j.PatternLayout
+log4j.appender.stdout.layout.ConversionPattern=%p %d{yyyy-MM-dd HH:mm:ss} [%c.%M(%L)] %m%n
+
+log4j.logger.com.heleeos.blog=INFO
+log4j.logger.com.heleeos.blog.dao=DEBUG
