--- Blog/Blog-manager/src/main/resources/log4j2.xml
@@ -1,16 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<Configuration status="WARN">
-    <Appenders>
-        <Console name="Console" target="SYSTEM_OUT">
-            <PatternLayout pattern="%p %d{yyyy-MM-dd HH:mm:ss} [%c.%M(%L)] %msg%n" />
-        </Console>
-    </Appenders>
-    <Loggers>
-        <Logger name="com.heleeos" level="info">
-            <AppenderRef ref="Console" />
-        </Logger>
-        <Root level="error">
-            <AppenderRef ref="Console" />
-        </Root>
-    </Loggers>
-</Configuration>
