+++ Blog/Blog-manager/src/main/webapp/WEB-INF/pages/blog/list.ftl
@@ -30,11 +30,11 @@
             <tbody>
                 <tr v-for="bean in beans">
                     <td>{{ bean.id }}</td>
-                    <td>{{ bean.title }}</td>
-                    <td>{{ bean.time | datetime }}</td>
-                    <td>{{ bean.lasttime | datetime }}</td>
+                    <td>{{ bean.blogTitle }}1</td>
+                    <td>{{ bean.createTime | datetime }}</td>
+                    <td>{{ bean.updateTime | datetime }}</td>
                     <td>{{ bean.dispIndex }}</td>
-                    <td>{{ dispState(bean.state) }}</td>
+                    <td>{{ dispState(bean.blogState) }}</td>
                     <td>
                         <button class="layui-btn layui-btn-primary" @click="editor(bean.id)">编辑</button>
                         <button class="layui-btn layui-btn-normal" @click="changeIndex(bean.id, 1)">上调</button>
@@ -60,32 +60,32 @@
                     beans: []
                 },
                 methods: {
-                	dispState: function(state) {
-                		switch (state) {
-							case 0: return "正常状态";
-							case 1: return "删除状态";
-							case 2: return "草稿状态";
-							default: return "其他状态";
-						}
-                	},                    
+                    dispState: function(state) {
+                        switch (state) {
+                            case 0: return "正常状态";
+                            case 1: return "删除状态";
+                            case 2: return "草稿状态";
+                            default: return "其他状态";
+                        }
+                    },                    
                     editor: function(id) {
                         parent.openDiv('', 'blog/add.html?id=' + id, loadBean);
                     },
                     changeIndex: function(id, change) {
-                    	$.post("changeIndex.json", { "id" : id, "change" : change}).done(function(res){
-                    		if(res.code == 200) {
-                    			parent.dispTip("修改成功");
-                    			loadBean();
-                    		} else {
-                    			parent.dispTip(res.message.info);
-                    		}
-                    	}).fail(function(){
-                    		parent.dispMessage("调整位置", "接口请求失败", true);
-                    	})
+                        $.post("/ajax/blog/changeIndex.json", { "id" : id, "change" : change}).done(function(res){
+                            if(res.code === 200) {
+                                parent.dispTip("修改成功");
+                                loadBean();
+                            } else {
+                                parent.dispTip(res.message.info);
+                            }
+                        }).fail(function(){
+                            parent.dispMessage("调整位置", "接口请求失败", true);
+                        })
                     },
                     changeState: function(id, state) {
-                    	$.post("changeState.json", { "id" : id, "state" : state}).done(function(res){
-                            if(res.code == 200) {
+                        $.post("/ajax/blog/changeState.json", { "id" : id, "state" : state}).done(function(res){
+                            if(res.code === 200) {
                                 parent.dispTip("修改成功");
                                 loadBean();
                             } else {
@@ -99,8 +99,8 @@
             });
             
             function loadBean() {
-            	$.post("list.json").done(function(res){
-                    if(res.code == 0) {
+                $.post("/ajax/blog/list.json").done(function(res){
+                    if(res.code === 0) {
                         vm.beans = res.message.beans;
                         if(vm.loadr) vm.loadr = false;
                     }
