+++ Blog/Blog-manager/src/main/webapp/WEB-INF/pages/main/logs.ftl
@@ -4,6 +4,10 @@
 </head>
 <body>
     <div class="layui-code">
+        Version 1.0.0 (2017-10-13)
+        1.重构所有的文件结构
+        2.重构博客bean和数据库
+        <br/>
         Version 0.1.9 (2017-4-13)
         1.[新增]文章上下调功能
         2.[完善]文章的状态编辑
